Refactor ProductFilling extension with controller
